// src/Momentum.Frontend/apps/code-innovation-backlog/src/generated/models/Cycai_votesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Cycai_votescycai_targettype = {
  100000000: 'Idea',
  100000001: 'Solution'
} as const;
export type Cycai_votescycai_targettype = keyof typeof Cycai_votescycai_targettype;
export const Cycai_votesstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Cycai_votesstatecode = keyof typeof Cycai_votesstatecode;
export const Cycai_votesstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Cycai_votesstatuscode = keyof typeof Cycai_votesstatuscode;

export interface Cycai_votesBase {
  cycai_name: string;
  cycai_targetid: number;
  cycai_targetkey: string;
  cycai_targettype: Cycai_votescycai_targettype;
  cycai_voteid: string;
  "cycai_voterid@odata.bind"?: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Cycai_votesstatecode;
  statuscode?: Cycai_votesstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Cycai_votes extends Cycai_votesBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  cycai_targettypename?: string;
  cycai_voteridname?: string;
  cycai_voteridyominame?: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  cycai_voterid?: object;
  _cycai_voterid_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}