// src/Momentum.Frontend/apps/code-innovation-backlog/src/generated/models/Cycai_activitiesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Cycai_activitiescycai_actortype = {
  100000000: 'User',
  100000001: 'Agent',
  100000002: 'System'
} as const;
export type Cycai_activitiescycai_actortype = keyof typeof Cycai_activitiescycai_actortype;
export const Cycai_activitiescycai_subjecttype = {
  100000000: 'Idea',
  100000001: 'Solution'
} as const;
export type Cycai_activitiescycai_subjecttype = keyof typeof Cycai_activitiescycai_subjecttype;
export const Cycai_activitiesstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Cycai_activitiesstatecode = keyof typeof Cycai_activitiesstatecode;
export const Cycai_activitiesstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Cycai_activitiesstatuscode = keyof typeof Cycai_activitiesstatuscode;

export interface Cycai_activitiesBase {
  cycai_action: string;
  cycai_activityid: string;
  "cycai_actorid@odata.bind"?: string;
  cycai_actortype?: Cycai_activitiescycai_actortype;
  cycai_name: string;
  cycai_occurredon?: string;
  cycai_subjectid?: number;
  cycai_subjectkey?: string;
  cycai_subjecttype?: Cycai_activitiescycai_subjecttype;
  cycai_summary?: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Cycai_activitiesstatecode;
  statuscode?: Cycai_activitiesstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Cycai_activities extends Cycai_activitiesBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  cycai_actoridname?: string;
  cycai_actoridyominame?: string;
  cycai_actortypename?: string;
  cycai_subjecttypename?: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  cycai_actorid?: object;
  _cycai_actorid_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}